Move k8s wrapper into a single package
