+++ src/main/java/com/microsoft/jenkins/kubernetes/credentials/ConfigFileCredentials.java
@@ -6,7 +6,7 @@
 
 package com.microsoft.jenkins.kubernetes.credentials;
 
-import com.microsoft.jenkins.kubernetes.KubernetesClientWrapper;
+import com.microsoft.jenkins.kubernetes.wrapper.KubernetesClientWrapper;
 import com.microsoft.jenkins.kubernetes.Messages;
 import hudson.Extension;
 import hudson.FilePath;
