+++ src/main/java/com/microsoft/jenkins/kubernetes/wrapper/ExtensionV1beta1ResourceUpdateMonitor.java
@@ -4,7 +4,7 @@
  * license information.
  */
 
-package com.microsoft.jenkins.kubernetes;
+package com.microsoft.jenkins.kubernetes.wrapper;
 
 import io.kubernetes.client.models.ExtensionsV1beta1Deployment;
