+++ src/main/java/com/microsoft/jenkins/kubernetes/wrapper/V1beta1ResourceUpdateMonitor.java
@@ -4,7 +4,7 @@
  * license information.
  */
 
-package com.microsoft.jenkins.kubernetes;
+package com.microsoft.jenkins.kubernetes.wrapper;
 
 import io.kubernetes.client.models.V1beta1DaemonSet;
 import io.kubernetes.client.models.V1beta1Ingress;
