+++ src/main/resources/com/microsoft/jenkins/kubernetes/Messages.properties
@@ -17,14 +17,10 @@ JobContext_nullContent = null content returned
 
 KubernetesClientWrapper_loadingConfiguration = Loading configuration: {0}
 KubernetesClientWrapper_noResourceLoadedFrom = No resource loaded from: {0}
-KubernetesClientWrapper_applied = Applied {0}: {1}
-KubernetesClientWrapper_created = Created {0}: {1}
 KubernetesClientWrapper_skipped = Skipped unsupported resource: {0}
 KubernetesClientWrapper_prepareSecretsWithName = Prepare Docker container registry secrets with name: {0}
 KubernetesClientWrapper_secretNameTooLong = ERROR: Secret name is longer than 253 characters: {0}
 KubernetesClientWrapper_illegalSecretName = ERROR: Illegal secret name: ''{0}''. See https://kubernetes.io/docs/concepts/overview/working-with-objects/names/ for reference.
-KubernetesClientWrapper_resourceNotFound = {0} (name: {1}) was not found in the Kubernetes cluster.
-KubernetesClientWrapper_noName = %s does not have name: %s
 
 DeploymentCommand_blankNamespace = Kubernetes secret namespace is not specified
 DeploymentCommand_blankConfigFiles = Kubernetes config files are not specified.
