+++ src/main/resources/com/microsoft/jenkins/kubernetes/wrapper/Messages.properties
@@ -1,0 +1,10 @@
+#
+# Copyright (c) Microsoft Corporation. All rights reserved.
+# Licensed under the MIT License. See License.txt in the project root for
+# license information.
+#
+
+KubernetesClientWrapper_noName = %s does not have name: %s
+KubernetesClientWrapper_applied = Applied {0}: {1}
+KubernetesClientWrapper_created = Created {0}: {1}
+KubernetesClientWrapper_resourceNotFound = {0} (name: {1}) was not found in the Kubernetes cluster.
